pyproject.toml
[tool.poetry]
name = "lesson-scheduler-api"
version = "0.1.1"
description = ""
authors = ["Mohammed Akmal Miah <mohammedmiah67890@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "3.13.1"
django-rest-auth = {extras = ["with_social"], version = "^0.9.5"}
django-cors-headers = "^3.11.0"
psycopg2 = "^2.9.3"
django-filter = "^21.1"
model-bakery = "^1.5.0"
drf-spectacular = "^0.22.1"
django-environ = "^0.8.1"
drf-extensions = "^0.7.1"
Django = "3.2.19"
drf-nested-routers = "^0.93.4"
mixpanel = "^4.10.0"
debugpy = "^1.6.6"
openai = "^0.27.2"
numpy = "^1.24.2"
python-dateutil = "^2.8.2"
django-rest-knox = "^4.2.0"


[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
include = '\.pyi?$'
exclude = [
  ".git",
  "migrations/",
  ".venv"
]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 119
default_section = "THIRDPARTY"
known_first_party = []
known_third_party = []

[tool.flake8]
max-line-length = 119
exclude = [
  ".git",
  "__pycache__",
  "build",
  "dist",
  "venv"
]
ignore = [
  "E203",
  "E501",
  "W503",
  "N807",
  "E712",
  "DJ01"
]
per-file-ignores = [
  # unused imports
  "__init__.py: F401"
]

[tool.bandit]
skips = ["B101"]